+++ src/App.css
@@ -28,7 +28,7 @@ h1 {
   grid-row: 2;
   /* background-color: rgb(189, 172, 172); */
 }
-.contentWrapper {
+.MainContentWrapper {
   grid-column: 1 / 3;
   grid-row: 3 / 10;
   /* background-color: green; */
