+++ src/components/Content/Content.css
@@ -1,0 +1,29 @@
+.wrapperContent {
+  display: grid;
+  grid-template-columns: 50% 50%;
+  grid-auto-rows: minmax(50px, auto);
+}
+
+.headline {
+  grid-column: 1;
+  grid-row: 2;
+  /* background-color: yellow;*/
+}
+
+.seeAll {
+  grid-column: 2;
+  grid-row: 2;
+
+  text-align: right;
+  padding-top: 25px;
+  padding-bottom: 25px;
+  color: grey;
+
+  /* background-color: rgb(252, 106, 106); */
+}
+
+.cards {
+  grid-column: 1 / 3;
+  grid-row: 3;
+  /* background-color: rgb(115, 221, 194); */
+}
