+++ src/components/Main/Main.css
@@ -1,4 +1,4 @@
-.wrapperContent {
+.wrapperMain {
   display: grid;
   grid-template-columns: 50% 50%;
   grid-auto-rows: minmax(50px, auto);
@@ -35,26 +35,7 @@
   border-bottom: 3px solid black;
 }
 
-.headline {
-  grid-column: 1;
-  grid-row: 2;
-  /* background-color: yellow;*/
-}
-
-.seeAll {
-  grid-column: 2;
-  grid-row: 2;
-
-  text-align: right;
-  padding-top: 25px;
-  padding-bottom: 25px;
-  color: grey;
-
-  /* background-color: rgb(252, 106, 106); */
-}
-
-.cards {
+.contentWrapper {
   grid-column: 1 / 3;
-  grid-row: 3;
-  /* background-color: rgb(115, 221, 194); */
+  grid-row: 1;
 }
