+++ src/data/data.js
@@ -1,0 +1,20 @@
+export var books = [
+  {
+    pic: "https://covers.oreillystatic.com/images/0636920020141/lrg.jpg",
+    headline: "Running Lean: Iterate form Plan A to a Plan That Works",
+    context:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Euismod in pellentesque massa placerat duis ultricies lacus sed."
+  },
+  {
+    pic: "https://images-na.ssl-images-amazon.com/images/I/71iTueU1pBL.jpg",
+    headline: "Hooked: how to build habit-forming products",
+    context:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Euismod in pellentesque massa placerat duis ultricies lacus sed."
+  },
+  {
+    pic: "https://images-na.ssl-images-amazon.com/images/I/41iRRYgWxYL.jpg",
+    headline: "Inspired: how to create tech products customers love",
+    context:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Euismod in pellentesque massa placerat duis ultricies lacus sed."
+  }
+];
